using Long instead of int for session id
Conflicts:
	kie-api/src/main/java/org/kie/api/openicar/KnowledgeService.java

+++ kie-api/src/main/java/org/kie/api/openicar/KnowledgeService.java
@@ -23,9 +23,9 @@ public interface KnowledgeService {
 
     void releaseStatefulKnowledgeSession();
 
-    int getStatefulKnowledgeSessionId(KieSession ksession);
+    Long getStatefulKnowledgeSessionId(KieSession ksession);
 
-    int getCurrentStatefulKnowledgeSessionId();
+    Long getCurrentStatefulKnowledgeSessionId();
 
     boolean isStarted();
